requirements.txt
logzero  # see github.com/metachris/logzero
black
pydocstyle
dnszone
blackhc.progress_bar
